# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurant-portal/restaurants/{restaurantId}/meals:
    get:
      operationId: getRestaurantMeals
      tags: [restaurant-portal-meals]
      summary: Get all meals/products for a restaurant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: isActive
          in: query
          schema:
            type: string
            enum: [all, true, false]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of meals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantMeal"
    post:
      operationId: createRestaurantMeal
      tags: [restaurant-portal-meals]
      summary: Create a new meal/product
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealBody"
      responses:
        "201":
          description: Meal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantMeal"
  /restaurant-portal/restaurants/{restaurantId}/meals/{mealId}:
    put:
      operationId: updateRestaurantMeal
      tags: [restaurant-portal-meals]
      summary: Update an existing meal/product
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: mealId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMealBody"
      responses:
        "200":
          description: Meal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantMeal"
    delete:
      operationId: deleteRestaurantMeal
      tags: [restaurant-portal-meals]
      summary: Delete a meal/product
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: mealId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Meal deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
  /restaurant-portal/restaurants/{restaurantId}/upcoming-meals:
    get:
      operationId: getRestaurantUpcomingMeals
      tags: [restaurant-portal-upcoming]
      summary: Get upcoming meals for preparation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: mealSlot
          in: query
          schema:
            type: string
            enum: [all, lunch, dinner]
        - name: status
          in: query
          schema:
            type: string
            enum: [all, scheduled, locked, preparing, delivered, cancelled]
      responses:
        "200":
          description: Upcoming meals data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpcomingMealsData"
  /restaurant-portal/restaurants/{restaurantId}/upcoming-meals/counts:
    get:
      operationId: getRestaurantMealCounts
      tags: [restaurant-portal-upcoming]
      summary: Get meal count summary for a date range
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RestaurantIdParam"
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Meal counts by date and slot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyMealCount"
...
  /student/auth/send-otp:
    post:
      operationId: studentSendOtp
      tags: [student-auth]
      summary: Send OTP to phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentSendOtpBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  devNote:
                    type: string
        "400":
          description: Invalid phone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /student/auth/verify-otp:
    post:
      operationId: studentVerifyOtp
      tags: [student-auth]
      summary: Verify OTP and get auth token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentVerifyOtpBody"
      responses:
        "200":
          description: Auth token and student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentAuthResponse"
        "401":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /student/restaurants:
    get:
      operationId: listStudentRestaurants
      tags: [student-restaurants]
      summary: List active restaurants
      security:
        - bearerAuth: []
      parameters:
        - name: campusId
          in: query
          schema:
            type: string
        - name: slot
          in: query
          schema:
            type: string
            enum: [lunch, dinner, both]
      responses:
        "200":
          description: Restaurant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentRestaurant"
  /student/restaurants/{restaurantId}:
    get:
      operationId: getStudentRestaurant
      tags: [student-restaurants]
      summary: Get restaurant detail with packages and meals
      security:
        - bearerAuth: []
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restaurant detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRestaurantDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RestaurantIdParam:
      name: restaurantId
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    RestaurantMeal:
      type: object
      properties:
        id:
          type: string
        restaurantId:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        category:
          type: string
        vegType:
          type: string
          enum: [veg, non-veg, egg]
        calories:
          type: integer
        protein:
          type: string
        carbs:
          type: string
        fat:
          type: string
        fiber:
          type: string
        allergens:
          type: string
        spiceLevel:
          type: string
          enum: [mild, medium, hot]
        imageUrl:
          type: string
        isAvailableForLunch:
          type: boolean
        isAvailableForDinner:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - restaurantId
        - name
        - vegType
        - isAvailableForLunch
        - isAvailableForDinner
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
    CreateMealBody:
      type: object
      properties:
        name:
          type: string
        shortDescription:
          type: string
        description:
          type: string
        category:
          type: string
        vegType:
          type: string
          enum: [veg, non-veg, egg]
        calories:
          type: integer
        protein:
          type: string
        carbs:
          type: string
        fat:
          type: string
        fiber:
          type: string
        allergens:
          type: string
        spiceLevel:
          type: string
          enum: [mild, medium, hot]
        imageUrl:
          type: string
        isAvailableForLunch:
          type: boolean
        isAvailableForDinner:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
      required:
        - name
    UpcomingMealsData:
      type: object
      properties:
        date:
          type: string
        lunchTotal:
          type: integer
        lunchLocked:
          type: integer
        lunchCancelled:
          type: integer
        lunchDelivered:
          type: integer
        dinnerTotal:
          type: integer
        dinnerLocked:
          type: integer
        dinnerCancelled:
          type: integer
        dinnerDelivered:
          type: integer
        isLockPassed:
          type: boolean
        lockTime:
          type: string
        orders:
          type: array
          items:
            $ref: "#/components/schemas/MealOrderDetail"
      required:
        - date
        - lunchTotal
        - lunchLocked
        - lunchCancelled
        - lunchDelivered
        - dinnerTotal
        - dinnerLocked
        - dinnerCancelled
        - dinnerDelivered
        - isLockPassed
        - lockTime
        - orders
    MealOrderDetail:
      type: object
      properties:
        id:
          type: string
        studentName:
          type: string
        studentPhoneMasked:
          type: string
        packageName:
          type: string
        mealSlot:
          type: string
          enum: [lunch, dinner]
        scheduledDate:
          type: string
        status:
          type: string
        freeCancelUntil:
          type: string
        isLocked:
          type: boolean
        pricePerDay:
          type: number
      required:
        - id
        - studentName
        - studentPhoneMasked
        - packageName
        - mealSlot
        - scheduledDate
        - status
        - freeCancelUntil
        - isLocked
        - pricePerDay
    DailyMealCount:
      type: object
      properties:
        date:
          type: string
        lunch:
          type: integer
        dinner:
          type: integer
        total:
          type: integer
        cancelled:
          type: integer
      required:
        - date
        - lunch
        - dinner
        - total
        - cancelled
...
    StudentSendOtpBody:
      type: object
      properties:
        phone:
          type: string
      required: [phone]
    StudentVerifyOtpBody:
      type: object
      properties:
        phone:
          type: string
        otp:
          type: string
      required: [phone, otp]
    StudentProfile:
      type: object
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: string
        campusId:
          type: string
          nullable: true
        campusName:
          type: string
          nullable: true
        foodPreference:
          type: string
          enum: [veg, non-veg, egg, jain]
        address:
          type: string
          nullable: true
        walletBalance:
          type: number
        isProfileComplete:
          type: boolean
        createdAt:
          type: string
      required: [id, phone, name, foodPreference, walletBalance, isProfileComplete]
    StudentAuthResponse:
      type: object
      properties:
        token:
          type: string
        student:
          $ref: "#/components/schemas/StudentProfile"
      required: [token, student]
...
    StudentRestaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tagline:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        cuisineType:
          type: string
          nullable: true
        campusId:
          type: string
          nullable: true
        isVeg:
          type: boolean
        lunchAvailable:
          type: boolean
        dinnerAvailable:
          type: boolean
        operatingDays:
          type: array
          items:
            type: string
        maxCapacity:
          type: integer
          nullable: true
        accentColor:
          type: string
          nullable: true
        lunchStartPrice:
          type: number
          nullable: true
        dinnerStartPrice:
          type: number
          nullable: true
        cancelCutoffHour:
          type: integer
        deliveryTime:
          type: string
          nullable: true
        distanceLabel:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        reviewCount:
          type: integer
      required: [id, name, isVeg, lunchAvailable, dinnerAvailable, operatingDays, cancelCutoffHour, reviewCount]
    StudentPackage:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        durationDays:
          type: integer
        mealSlot:
          type: string
          enum: [lunch, dinner, both]
        pricePerDay:
          type: number
        totalPrice:
          type: number
        discountPercent:
          type: number
        mealsPerDay:
          type: integer
      required: [id, name, durationDays, mealSlot, pricePerDay, totalPrice, discountPercent, mealsPerDay]
    StudentMeal:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        shortDescription:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        vegType:
          type: string
          enum: [veg, non-veg, egg]
        spiceLevel:
          type: string
          enum: [mild, medium, hot]
          nullable: true
        calories:
          type: integer
          nullable: true
        storagePath:
          type: string
          nullable: true
        lunchAvailable:
          type: boolean
        dinnerAvailable:
          type: boolean
      required: [id, name, vegType, lunchAvailable, dinnerAvailable]
    StudentRestaurantDetail:
      allOf:
        - $ref: "#/components/schemas/StudentRestaurant"
        - type: object
          properties:
            packages:
              type: array
              items:
                $ref: "#/components/schemas/StudentPackage"
            meals:
              type: array
              items:
                $ref: "#/components/schemas/StudentMeal"
            ratingsBreakdown:
              type: object
              nullable: true
              properties:
                totalReviews:
                  type: integer
                overall:
                  type: number
                  nullable: true
                foodQuality:
                  type: number
                  nullable: true
                packaging:
                  type: number
                  nullable: true
                delivery:
                  type: number
                  nullable: true
                valueForMoney:
                  type: number
                  nullable: true
                hygiene:
                  type: number
                  nullable: true
          required: [packages, meals]
...